Possible Fix for #128
Entities outside of EntityPlayer were being treated as a Resident when passing "onLivingAttack." This is a small patch to check to see if ev.source.getEntity() (which was the reson for the crash since it was checking the entity and not the entity player itself) is an instanceof EntityPlayer. Also added a small null check for hasResident (could be removed if it does cause any issues).